pyproject.toml
[project]
name = "ml-review-env"
version = "1.0.0"
description = "OpenEnv environment for ML training code review"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "pyyaml",
    "python-dotenv",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server:main"

[project.urls]
Homepage = "https://huggingface.co/spaces/raghava2/ml-review-env"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[tool.openenv]
name = "ml-code-review-env"
version = "1.0.0"
tasks = ["broken_training_loop", "broken_validation_loop", "broken_transformer"]
action_space = ["submit_fix", "explain_issue", "optimize", "request_hint", "no_op"]
reward_range = [-0.2, 1.0]